Starting to move fhd tests to separate file
